refactor: move db backup code into service
