+++ server/src/services/index.ts
@@ -7,7 +7,6 @@ import { AssetService } from 'src/services/asset.service';
 import { AuditService } from 'src/services/audit.service';
 import { AuthAdminService } from 'src/services/auth-admin.service';
 import { AuthService } from 'src/services/auth.service';
-import { BackupService } from 'src/services/backup.service';
 import { CliService } from 'src/services/cli.service';
 import { DatabaseBackupService } from 'src/services/database-backup.service';
 import { DatabaseService } from 'src/services/database.service';
@@ -58,7 +57,6 @@ export const services = [
   AuditService,
   AuthService,
   AuthAdminService,
-  BackupService,
   CliService,
   DatabaseBackupService,
   DatabaseService,
